# krishi_connect/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Farmer Platform API",
        "description": "API for farmers in Maharashtra with market rates, schemes, and AI advisory.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register User",
                "operationId": "register_user_api_v1_auth_register_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login For Access Token",
                "operationId": "login_for_access_token_api_v1_auth_login_post",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_login_for_access_token_api_v1_auth_login_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Read Users Me",
                "operationId": "read_users_me_api_v1_auth_me_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
        "/api/v1/farmer/status": {
            "get": {
                "tags": [
                    "Farmer"
                ],
                "summary": "Farmer Status",
                "operationId": "farmer_status_api_v1_farmer_status_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/farmer/crops": {
            "get": {
                "tags": [
                    "Farmer"
                ],
                "summary": "List Crops",
                "description": "List all available crops for advisory and market rates.",
                "operationId": "list_crops_api_v1_farmer_crops_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Crop"
                                    },
                                    "title": "Response List Crops Api V1 Farmer Crops Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/farmer/schemes": {
            "get": {
                "tags": [
                    "Farmer"
                ],
                "summary": "List Schemes",
                "description": "List all government schemes.",
                "operationId": "list_schemes_api_v1_farmer_schemes_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GovernmentScheme"
                                    },
                                    "title": "Response List Schemes Api V1 Farmer Schemes Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/farmer/advisory/soil": {
            "post": {
                "tags": [
                    "Farmer"
                ],
                "summary": "Get Soil And Crop Advisory",
                "description": "Get AI-powered soil and crop advisory based on farmer input.",
                "operationId": "get_soil_and_crop_advisory_api_v1_farmer_advisory_soil_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SoilAdvisoryRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SoilAdvisoryResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/farmer/chatbot": {
            "post": {
                "tags": [
                    "Farmer"
                ],
                "summary": "Chatbot Query",
                "description": "Send a message to the AI Chatbot and get a response with updated history.",
                "operationId": "chatbot_query_api_v1_farmer_chatbot_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatbotMessage"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatbotResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/farmer/market_rates": {
            "get": {
                "tags": [
                    "Farmer"
                ],
                "summary": "List Market Rates",
                "description": "List daily market rates, with optional filters.",
                "operationId": "list_market_rates_api_v1_farmer_market_rates_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "crop_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "title": "Crop Id"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Location"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Date"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MarketRate"
                                    },
                                    "title": "Response List Market Rates Api V1 Farmer Market Rates Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_login_for_access_token_api_v1_auth_login_post": {
                "properties": {
                    "grant_type": {
                        "anyOf": [
                            {
                                "type": "string",
                                "pattern": "^password$"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Grant Type"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "title": "Password"
                    },
                    "scope": {
                        "type": "string",
                        "title": "Scope",
                        "default": ""
                    },
                    "client_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Client Id"
                    },
                    "client_secret": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "format": "password",
                        "title": "Client Secret"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "password"
                ],
                "title": "Body_login_for_access_token_api_v1_auth_login_post"
            },
...
            "ChatbotMessage": {
                "properties": {
                    "session_id": {
                        "type": "string",
                        "title": "Session Id"
                    },
                    "message_content": {
                        "type": "string",
                        "title": "Message Content"
                    }
                },
                "type": "object",
                "required": [
                    "session_id",
                    "message_content"
                ],
                "title": "ChatbotMessage"
            },
            "ChatbotResponse": {
                "properties": {
                    "response": {
                        "type": "string",
                        "title": "Response"
                    },
                    "history": {
                        "items": {
                            "$ref": "#/components/schemas/ChatbotHistoryItem"
                        },
                        "type": "array",
                        "title": "History"
                    }
                },
                "type": "object",
                "required": [
                    "response",
                    "history"
                ],
                "title": "ChatbotResponse"
            },
            "Crop": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "title": "Name"
                    },
                    "image_url": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 512
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Url"
                    },
                    "details": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Details"
                    },
                    "context": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Context"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "id"
                ],
                "title": "Crop"
            },
...
            "GovernmentScheme": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 512,
                        "title": "Title"
                    },
                    "link": {
                        "type": "string",
                        "maxLength": 512,
                        "title": "Link"
                    },
                    "image_url": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 512
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Image Url"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "region": {
                        "type": "string",
                        "maxLength": 255,
                        "title": "Region",
                        "default": "Maharashtra"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "created_by_id": {
                        "type": "integer",
                        "title": "Created By Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "link",
                    "id",
                    "created_by_id",
                    "created_at"
                ],
                "title": "GovernmentScheme"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "SoilAdvisoryRequest": {
                "properties": {
                    "crop_id": {
                        "type": "integer",
                        "title": "Crop Id",
                        "description": "ID of the selected crop."
                    },
                    "soil_type": {
                        "type": "string",
                        "title": "Soil Type",
                        "description": "e.g., Black soil, Red soil, Alluvial soil."
                    },
                    "ph_level": {
                        "type": "number",
                        "title": "Ph Level",
                        "description": "Soil pH level."
                    },
                    "nutrient_levels": {
                        "type": "string",
                        "title": "Nutrient Levels",
                        "description": "e.g., NPK values, Organic Carbon content."
                    },
                    "weather_conditions": {
                        "type": "string",
                        "title": "Weather Conditions",
                        "description": "Current or expected weather conditions."
                    }
                },
                "type": "object",
                "required": [
                    "crop_id",
                    "soil_type",
                    "ph_level",
                    "nutrient_levels",
                    "weather_conditions"
                ],
                "title": "SoilAdvisoryRequest"
            },
            "SoilAdvisoryResponse": {
                "properties": {
                    "crop_name": {
                        "type": "string",
                        "title": "Crop Name"
                    },
                    "advisory": {
                        "type": "string",
                        "title": "Advisory",
                        "description": "Detailed, localized advisory from the AI."
                    }
                },
                "type": "object",
                "required": [
                    "crop_name",
                    "advisory"
                ],
                "title": "SoilAdvisoryResponse"
            },
            "Token": {
                "properties": {
                    "access_token": {
                        "type": "string",
                        "title": "Access Token"
                    },
                    "token_type": {
                        "type": "string",
                        "title": "Token Type",
                        "default": "bearer"
                    }
                },
                "type": "object",
                "required": [
                    "access_token"
                ],
                "title": "Token"
            },
            "User": {
                "properties": {
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "role": {
                        "type": "string",
                        "pattern": "^(farmer|admin)$",
                        "title": "Role"
                    },
                    "id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Id"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active",
                        "default": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "role",
                    "created_at"
                ],
                "title": "User"
            },
            "UserCreate": {
                "properties": {
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "role": {
                        "type": "string",
                        "pattern": "^(farmer|admin)$",
                        "title": "Role"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "role",
                    "password"
                ],
                "title": "UserCreate"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/v1/auth/login"
                    }
                }
            }
        }
    }
}